Addressed @sbc's comments Split tests/audioworklet into a futex test (without the tone generator) and a tone generator example (without the test stuff) Fixed `moduleLoaded` failing (it needs to register on the global scope) Made everything build with MINIMAL_RUNTIME (doesn't run just yet) Removed the acron fix temporarily
